+++ core/src/com/color/game/elements/staticelements/sensors/Sensor.java
@@ -20,7 +20,7 @@ public abstract class Sensor extends BaseStaticElement {
      * @param height the height of the Sensor
      * @param map the map containing the Sensor
      */
-    Sensor(Vector2 position, int width, int height, Map map) {
+    Sensor(Vector2 position, float width, float height, Map map) {
         super(position, width, height, map, PhysicComponent.CATEGORY_SENSOR, PhysicComponent.MASK_SENSOR);
         this.physicComponent.configureUserData(new StaticElementUserData(this, width, height, UserDataType.SENSOR));
     }
