+++ core/src/com/color/game/elements/userData/DynamicPlatformUserData.java
@@ -4,7 +4,7 @@ import com.color.game.elements.dynamicplatforms.BaseDynamicPlatform;
 
 public class DynamicPlatformUserData extends UserData {
 
-    public DynamicPlatformUserData(BaseDynamicPlatform element, int width, int height, UserDataType userDataType) {
+    public DynamicPlatformUserData(BaseDynamicPlatform element, float width, float height, UserDataType userDataType) {
         super(element, width, height, userDataType);
     }
 }
